requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lightning-utilities==0.14.2
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.2.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
torchmetrics==1.7.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
yarl==1.18.3
